backend/requirements.txt
flask
flask-cors
scikit-learn
sentence-transformers
pypdf
python-docx
numpy
